requirements.txt
Flask
configparser
mysql.connector
